Adjusted Movement and Idle Animation and Added Player Corpses
Changed idle animation and fixed it to make the body move up and down instead of moving the body and vine "feet" away from each other, which was making the animation look weird and causing the "feet" to go inside of the ground.

Made adjustments to the effects of gravity on and friction of the player as well as the movement and jump forces to improve the feel of the movement. Added a !_isGrounded check to the wallJump horizontal force check so that you are not pushed back when jumping against a wall from the ground. Movement overall needs much more fine tuning.

Also changed ground and wall colliders from triggers on the environment layer to non-triggers on separate "Ground Box" and "Wall Box" layers, which do not interact with each other or the environment, player, or default layers in the Physics2D matrix so as to essentially only interact with the player's grounded and wall sliding rays. This was to fix a discrepancy issue with the rays between the physics collider and ground/wall colliders, which was causing the player to rapidly toggle the grounded and wall sliding states.

Used player game object as a basis to create a player corpse prefab. All of the player-related components, tags, and layers were removed/changed. Created a new sprite for the corpse and a PlayerCorpse script to count how many corpses have spawned after the corpse with the script attached, and destroy the corpse when that count passes a set maximum number of corpses. This creates a system where the player can have up to 3 corpses at once, and creating a new corpse replaces the oldest one. Corpses utilize a new static EventManager class to run a custom onCorpseSpawned event.
EventManager also includes onPlayerSpawned and onPlayerDied events. onPlayerSpawned isused by the new player camera object + script to reset the follow target. onPlayerDied is used by the new GameManager object + script to spawn a new player game object after a short delay and a new death counter UI object + script to update the text on the death counter.

TODO: Create hazards that can kill the player, then start level creation. Make a more extensive gameplay UI and some tutorial instructions. Add running, jumping, falling, wall sliding, wall jumping, and death animations. Make a main menu and level selection screen. Make art/animations for a checkpoint, hazards, platforms, UI, and background(s).